refactor(server): extract dashboard static plugin (#3154)
Phase 1, extraction #3: Dashboard static file serving.

Moves 175 lines of dashboard serving code from server.ts to a new
Fastify plugin at src/plugins/dashboard-static.ts:

- CSP headers, cache control, static file serving
- SPA fallback for dashboard routes
- Root redirect to /dashboard/
- manifest.json for PWA install
- Content-Length defensive header

server.ts: 1513 → 1384 lines (-129 lines, -8.5%)
No API changes. All dashboard behavior preserved.
